+++ foods.html
@@ -9,14 +9,11 @@
     <thead class="food-table-head">
       <th>Name</th>
       <th>Calories</th>
-      <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
       <script src="main.bundle.js"></script>
     </thead>
     <tr>
     </tr>
   </table>
-<<<<<<< HEAD
-=======
 
   <div class="food-form">
     <h3>Manage Foods</h3>
@@ -28,14 +25,5 @@
       <button id="add_food">Add Food</button>
     </form>
   </div>
-
-
-  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
-  <script src="main.bundle.js"></script>
->>>>>>> 6f598d8e3486025cf6cef7fcb8c73c67827915c5
 </body>
 </html>
-
-
-<!-- When I visit foods.html, I can enter a name and calorie amount, and create a new
-food by clicking "Add Food" -->
